+++ src/main/java/life/majiang/community/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable Integer id,
+    public String question(@PathVariable Long id,
                            Model model) {
         QuestionDTO questionDTO = questionService.findById(id);
         questionService.incView(id);
